Move the tests module to a separate source file
Break things up so the individual files are shorter and save
an indent level.
